requirements.txt
fastapi>=0.110.1,<0.111.1
pydantic-settings>=2.2.1,<2.3.0
uvicorn[standard]>=0.29.0,<0.30.0
black>=24.3.0,<24.4.0
geojson-pydantic>=1.0.1,<1.0.2
rio-tiler>=6.6.1,<6.7.0
rasterio>=1.3.9,<1.3.10
geopandas>=1.0.0,<1.0.1
shapely>=2.0.3,<2.0.4
pyjwt>=2.7.0,<3.0.0
python-jose>=3.3.0,<4.0.0